+++ src/TehPers.FishingOverhaul.Api/Content/FishEntry.cs
@@ -1,7 +1,6 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 using Newtonsoft.Json;
 using TehPers.Core.Api.Items;
-using TehPers.Core.Api.Json;
 
 namespace TehPers.FishingOverhaul.Api.Content
 {
@@ -10,7 +9,6 @@ namespace TehPers.FishingOverhaul.Api.Content
     /// </summary>
     /// <param name="FishKey">The item key.</param>
     /// <param name="AvailabilityInfo">The availability information.</param>
-    [JsonDescribe]
     public record FishEntry(
         [property: JsonRequired] NamespacedKey FishKey,
         FishAvailabilityInfo AvailabilityInfo