+++ src/TehPers.FishingOverhaul.Api/Content/FishEntryFilter.cs
@@ -1,12 +1,10 @@
 using TehPers.Core.Api.Items;
-using TehPers.Core.Api.Json;
 
 namespace TehPers.FishingOverhaul.Api.Content
 {
     /// <summary>
     /// Fish entry filter.
     /// </summary>
-    [JsonDescribe]
     public record FishEntryFilter
     {
         /// <summary>