+++ src/TehPers.FishingOverhaul.Api/Content/TrashEntryFilter.cs
@@ -1,12 +1,10 @@
 using TehPers.Core.Api.Items;
-using TehPers.Core.Api.Json;
 
 namespace TehPers.FishingOverhaul.Api.Content
 {
     /// <summary>
     /// Trash entry filter.
     /// </summary>
-    [JsonDescribe]
     public record TrashEntryFilter
     {
         /// <summary>