+++ src/TehPers.FishingOverhaul/Config/ContentPacks/TreasurePack.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Immutable;
-using TehPers.Core.Api.Json;
 using TehPers.FishingOverhaul.Api.Content;
 
 namespace TehPers.FishingOverhaul.Config.ContentPacks
@@ -7,13 +6,17 @@ namespace TehPers.FishingOverhaul.Config.ContentPacks
     /// <summary>
     /// Content which controls what treasure are available to catch.
     /// </summary>
-    [JsonDescribe]
     public record TreasurePack : JsonConfigRoot
     {
+        /// <inheritdoc cref="JsonConfigRoot.Schema" />
+        protected override string Schema =>
+            $"{JsonConfigRoot.jsonSchemaRootUrl}contentPacks/treasure.schema.json";
+
         /// <summary>
         /// The treasure entries to add.
         /// </summary>
-        public ImmutableArray<TreasureEntry> Add { get; init; } = ImmutableArray<TreasureEntry>.Empty;
+        public ImmutableArray<TreasureEntry> Add { get; init; } =
+            ImmutableArray<TreasureEntry>.Empty;
 
         /// <summary>
         /// Merges all the treasure entries into a single content object.
@@ -21,7 +24,7 @@ namespace TehPers.FishingOverhaul.Config.ContentPacks
         /// <param name="content">The content to merge into.</param>
         public FishingContent AddTo(FishingContent content)
         {
-            return content with { AddTreasure = content.AddTreasure.AddRange(this.Add) };
+            return content with {AddTreasure = content.AddTreasure.AddRange(this.Add)};
         }
     }
 }