+++ src/TehPers.FishingOverhaul/Parsing/UnaryOperator.cs
@@ -1,0 +1,13 @@
+﻿namespace TehPers.FishingOverhaul.Parsing
+{
+    /// <summary>
+    /// A unary operator in an expression.
+    /// </summary>
+    internal enum UnaryOperator
+    {
+        /// <summary>
+        /// The negation operator (<c>-</c>).
+        /// </summary>
+        Negate,
+    }
+}