+++ src/TehPers.FishingOverhaul/Services/DefaultFishingSource.FishData.cs
@@ -12,18 +12,22 @@ namespace TehPers.FishingOverhaul.Services
 {
     internal sealed partial class DefaultFishingSource
     {
-        private static readonly ImmutableDictionary<string, string> hasCaughtFish =
-            new Dictionary<string, string>
+        private static readonly ImmutableDictionary<string, string?> hasCaughtFish =
+            new Dictionary<string, string?>
             {
                 ["HasValue:{{HasCaughtFish}}"] = "true",
             }.ToImmutableDictionary();
 
-        private static readonly ImmutableDictionary<string, string> isLegendaryFamilyActive =
-            new Dictionary<string, string>
+        private static readonly ImmutableDictionary<string, string?> legendaryBaseConditions =
+            DefaultFishingSource.hasCaughtFish.Add(
+                "TehPers.FishingOverhaul/SpecialOrderRuleActive |contains=LEGENDARY_FAMILY",
+                "false"
+            );
+
+        private static readonly ImmutableDictionary<string, string?> isLegendaryFamilyActive =
+            new Dictionary<string, string?>
             {
                 ["TehPers.FishingOverhaul/SpecialOrderRuleActive"] = "LEGENDARY_FAMILY",
-                // TODO: remove this once CP updates
-                ["HasMod"] = "TehPers.FishingOverhaul",
             }.ToImmutableDictionary();
 
         // Legendary fish
@@ -45,8 +49,8 @@ namespace TehPers.FishingOverhaul.Services
                         GreaterThanEq = 82,
                     },
                 },
-                MinDepth = 3,
-                When = DefaultFishingSource.hasCaughtFish.Add(
+                MinBobberDepth = 3,
+                When = DefaultFishingSource.legendaryBaseConditions.Add(
                     $"HasFlag |contains={DefaultFishingSource.crimsonfishFlag}",
                     "false"
                 ),
@@ -77,7 +81,7 @@ namespace TehPers.FishingOverhaul.Services
                         LessThan = 15,
                     },
                 },
-                When = DefaultFishingSource.hasCaughtFish.Add(
+                When = DefaultFishingSource.legendaryBaseConditions.Add(
                     $"HasFlag |contains={DefaultFishingSource.anglerFlag}",
                     "false"
                 ),
@@ -103,8 +107,8 @@ namespace TehPers.FishingOverhaul.Services
                 Weathers = Weathers.Rainy,
                 WaterTypes = WaterTypes.PondOrOcean,
                 MinFishingLevel = 10,
-                MinDepth = 4,
-                When = DefaultFishingSource.hasCaughtFish.Add(
+                MinBobberDepth = 4,
+                When = DefaultFishingSource.legendaryBaseConditions.Add(
                     $"HasFlag |contains={DefaultFishingSource.legendFlag}",
                     "false"
                 ),
@@ -137,7 +141,7 @@ namespace TehPers.FishingOverhaul.Services
                         LessThanEq = 42,
                     },
                 },
-                When = DefaultFishingSource.hasCaughtFish.Add(
+                When = DefaultFishingSource.legendaryBaseConditions.Add(
                     $"HasFlag |contains={DefaultFishingSource.mutantCarpFlag}",
                     "false"
                 ),
@@ -175,8 +179,8 @@ namespace TehPers.FishingOverhaul.Services
                         LessThan = 88,
                     },
                 },
-                MinDepth = 3,
-                When = DefaultFishingSource.hasCaughtFish.Add(
+                MinBobberDepth = 3,
+                When = DefaultFishingSource.legendaryBaseConditions.Add(
                     $"HasFlag |contains={DefaultFishingSource.glacierfishFlag}",
                     "false"
                 ),
@@ -202,6 +206,8 @@ namespace TehPers.FishingOverhaul.Services
                 AvailabilityInfo =
                 DefaultFishingSource.crimsonfishEntry.AvailabilityInfo with
                 {
+                    Seasons = Seasons.All,
+                    Weathers = Weathers.All,
                     When = DefaultFishingSource.isLegendaryFamilyActive,
                 },
                 OnCatch = new()
@@ -221,6 +227,8 @@ namespace TehPers.FishingOverhaul.Services
             AvailabilityInfo =
             DefaultFishingSource.anglerEntry.AvailabilityInfo with
             {
+                Seasons = Seasons.All,
+                Weathers = Weathers.All,
                 When = DefaultFishingSource.isLegendaryFamilyActive,
             },
             OnCatch = new()
@@ -238,6 +246,8 @@ namespace TehPers.FishingOverhaul.Services
             AvailabilityInfo =
             DefaultFishingSource.legendEntry.AvailabilityInfo with
             {
+                Seasons = Seasons.All,
+                Weathers = Weathers.All,
                 When = DefaultFishingSource.isLegendaryFamilyActive,
             },
             OnCatch = new()
@@ -256,6 +266,8 @@ namespace TehPers.FishingOverhaul.Services
                 AvailabilityInfo =
                 DefaultFishingSource.mutantCarpEntry.AvailabilityInfo with
                 {
+                    Seasons = Seasons.All,
+                    Weathers = Weathers.All,
                     When = DefaultFishingSource.isLegendaryFamilyActive,
                 },
                 OnCatch = new()
@@ -275,6 +287,8 @@ namespace TehPers.FishingOverhaul.Services
                 AvailabilityInfo =
                 DefaultFishingSource.glacierfishEntry.AvailabilityInfo with
                 {
+                    Seasons = Seasons.All,
+                    Weathers = Weathers.All,
                     When = DefaultFishingSource.isLegendaryFamilyActive,
                 },
                 OnCatch = new()
@@ -309,7 +323,7 @@ namespace TehPers.FishingOverhaul.Services
             var trashEntries = new List<TrashEntry>();
 
             // Parse the fish traits
-            var fish = this.assetProvider.Load<Dictionary<int, string>>(@"Data\Fish.xnb");
+            var fish = this.assetProvider.Load<Dictionary<int, string>>(@"Data\Fish");
             var fishAvailabilities = new Dictionary<int, List<FishAvailabilityInfo>>(fish.Count);
             var trashAvailabilities = new Dictionary<int, List<AvailabilityInfo>>();
             foreach (var (fishId, rawFishInfo) in fish)
@@ -332,11 +346,15 @@ namespace TehPers.FishingOverhaul.Services
                     }
 
                     // Add traits and availabilities
-                    fishTraits[fishKey] = traits with
-                    {
-                        IsLegendary = DefaultFishingSource.vanillaLegendaries.Contains(fishKey),
-                    };
-                    fishAvailabilities[fishId] = availabilities.ToList();
+                    fishTraits.Add(
+                        fishKey,
+                        traits with
+                        {
+                            IsLegendary =
+                            DefaultFishingSource.vanillaLegendaries.Contains(fishKey),
+                        }
+                    );
+                    fishAvailabilities.Add(fishId, availabilities.ToList());
                 }
                 else
                 {
@@ -356,8 +374,7 @@ namespace TehPers.FishingOverhaul.Services
             }
 
             // Parse the location data
-            var locations =
-                this.assetProvider.Load<Dictionary<string, string>>(@"Data\Locations.xnb");
+            var locations = this.assetProvider.Load<Dictionary<string, string>>(@"Data\Locations");
             foreach (var (locationName, rawLocationData) in locations)
             {
                 var locationData = rawLocationData.Split('/');
@@ -679,7 +696,7 @@ namespace TehPers.FishingOverhaul.Services
                 availability => availability with
                 {
                     DepthMultiplier = depthMultiplier,
-                    MaxDepth = maxDepth
+                    MaxChanceDepth = maxDepth
                 }
             );