+++ src/TehPers.FishingOverhaul/Services/DefaultFishingSource.cs
@@ -9,6 +9,10 @@ namespace TehPers.FishingOverhaul.Services
 {
     internal sealed partial class DefaultFishingSource : IFishingContentSource
     {
+        private const double prioritizedTier = 10d;
+        private const double specialItemTier = 20d;
+        private const double questItemTier = 30d;
+
         private readonly IManifest manifest;
         private readonly IAssetProvider assetProvider;
 
@@ -33,6 +37,7 @@ namespace TehPers.FishingOverhaul.Services
             this.defaultContent.Add(this.GetDefaultFishData());
             this.defaultContent.Add(this.GetDefaultTrashData());
             this.defaultContent.Add(this.GetDefaultTreasureData());
+            this.defaultContent.Add(this.GetDefaultEffectData());
 
             return this.defaultContent;
         }