+++ Content Packs/[TFO] TehPers.RecatchableLegendaries/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog - Recatchable Legendaries
 
+## 1.0.2 - Unreleased
+
+### Changed
+
+- Updated to new content pack schema (`MinDepth` -> `MinBobberDepth`).
+
+### Fixed
+
+- Fix conditions for catching the legendary/legendary II fish.
+
 ## 1.0.1 - 2021-12-23
 
 ### Fixed